EPI-467: Add configurable validation to display ID (#4095)
* Add new localisation entry

* Extend patientDetailValidations to take extra fields

* Extend getLocalisedSchema to include regex validation if configured

* Include validation from config - only if edit option is turned on

* Rename DisplayIdField to DisplayIdSearchField

* Expect new pattern key in test

* Revert "Extend patientDetailValidations to take extra fields"

This reverts commit ad25734088d70f5f1aade265d4e2ee5da70e419f.

* Revert "Extend getLocalisedSchema to include regex validation if configured"

This reverts commit d0a025b199821cd498f82ab2812ecff203340606.

* Revert "Include validation from config - only if edit option is turned on"

This reverts commit 44a41e787ea1670c6d84883c0b03be1cad28412d.

* Create new DisplayIdField with special validation

* Use new DisplayIdField in form

* Make sure regex exists before testing against it


* Show error notification

---------
