+++ packages/desktop/app/components/SearchBar/CovidPatientsSearchBar.js
@@ -4,7 +4,7 @@ import {
   AutocompleteField,
   Field,
   LocalisedField,
-  DisplayIdField,
+  DisplayIdSearchField,
   DOBFields,
   SearchField,
 } from '../Field';
@@ -23,7 +23,7 @@ export const CovidPatientsSearchBar = React.memo(({ onSearch }) => {
       <LocalisedField name="firstName" component={SearchField} />
       <LocalisedField name="lastName" component={SearchField} />
       <LocalisedField name="villageId" component={AutocompleteField} suggester={villageSuggester} />
-      <DisplayIdField />
+      <DisplayIdSearchField />
       <DOBFields />
       <Field name="clinicalStatus" label="Clinical status" component={SearchField} />
       <SearchBarCheckField name="deceased" label="Include deceased patients" />
