+++ packages/sync-server/__tests__/middleware/auth.test.js
@@ -137,6 +137,7 @@ describe('Auth', () => {
       expect(response.body).toHaveProperty('localisation.fields.displayId', {
         shortLabel: 'NHN',
         longLabel: 'National Health Number',
+        pattern: '[\\s\\S]*',
       });
     });
