+++ packages/sync-server/app/localisation.js
@@ -17,6 +17,7 @@ const fieldSchema = yup
     }),
     hidden: yup.boolean().required(),
     required: yup.boolean(),
+    pattern: yup.string(),
   })
   .default({}) // necessary to stop yup throwing hard-to-debug errors
   .required()
@@ -27,6 +28,7 @@ const unhideableFieldSchema = yup
     shortLabel: yup.string().required(),
     longLabel: yup.string().required(),
     required: yup.boolean(),
+    pattern: yup.string(),
   })
   .required()
   .noUnknown();
